backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-multipart==0.0.6
google-auth==2.23.4
pydantic==2.5.0
python-dotenv==1.0.0